- Added points system - Prevented vote on self

+++ src/main/java/org/weviewapp/service/impl/VoteServiceImpl.java
@@ -14,6 +14,7 @@ import org.weviewapp.repository.CommentRepository;
 import org.weviewapp.repository.ReviewRepository;
 import org.weviewapp.repository.UserRepository;
 import org.weviewapp.repository.VoteRepository;
+import org.weviewapp.service.UserService;
 import org.weviewapp.service.VoteService;
 
 import java.util.Optional;
@@ -29,6 +30,8 @@ public class VoteServiceImpl implements VoteService {
     private CommentRepository commentRepository;
     @Autowired
     private UserRepository userRepository;
+    @Autowired
+    private UserService userService;
 
     public Vote vote(VoteOn voteOn, UUID id, UUID userId, VoteType voteType) {
         User user = userRepository.findById(userId)
@@ -48,14 +51,43 @@ public class VoteServiceImpl implements VoteService {
         Comment comment = commentRepository.findById(commentId)
                 .orElseThrow(() -> new WeviewAPIException(HttpStatus.BAD_REQUEST, "Comment not found!"));
 
+        if (comment.getUser().getId().equals(user.getId())) {
+            throw new WeviewAPIException(HttpStatus.FORBIDDEN, "Cannot vote on own comment!");
+        }
+
         Optional<Vote> existingVote = voteRepository.findByComment_IdAndUserId(comment.getId(), user.getId());
         if (existingVote.isPresent()) {
             if (existingVote.get().getVoteType().equals(voteType)) {
                 voteRepository.delete(existingVote.get());
+
+                // Point change to author
+                if (voteType.equals(VoteType.DOWNVOTE)) {
+                    // REMOVE DOWNVOTE
+                    userService.modifyPoints(comment.getUser().getId(), 2);
+                } else if (voteType.equals(VoteType.UPVOTE)) {
+                    //REMOVE UPVOTE
+                    userService.modifyPoints(comment.getUser().getId(), -2);
+                }
+
+                // Point change to current user
+                // Removing vote, so remove points
+                userService.modifyPoints(user.getId(), -1);
                 return null;
             } else {
                 existingVote.get().setVoteType(voteType);
-                return voteRepository.save(existingVote.get());
+                Vote vote = voteRepository.save(existingVote.get());
+
+                // Inverting votes
+                if (voteType.equals(VoteType.DOWNVOTE)) {
+                    // REMOVE UPVOTE ADD DOWNVOTE
+                    userService.modifyPoints(comment.getUser().getId(), -4);
+                } else if (voteType.equals(VoteType.UPVOTE)) {
+                    // REMOVE DOWNVOTE ADD UPVOTE
+                    userService.modifyPoints(comment.getUser().getId(), +4);
+                }
+
+                // Invert points so no changes in current user points
+                return vote;
             }
         } else {
             Vote vote = new Vote();
@@ -63,7 +95,19 @@ public class VoteServiceImpl implements VoteService {
             vote.setComment(comment);
             vote.setVoteType(voteType);
             vote.setUser(user);
-            return voteRepository.save(vote);
+            Vote newVote = voteRepository.save(vote);
+
+            if (voteType.equals(VoteType.DOWNVOTE)) {
+                // ADD DOWNVOTE
+                userService.modifyPoints(comment.getUser().getId(), -2);
+            } else if (voteType.equals(VoteType.UPVOTE)) {
+                // ADD UPVOTE
+                userService.modifyPoints(comment.getUser().getId(), +2);
+            }
+
+            // +1 for action
+            userService.modifyPoints(user.getId(), 1);
+            return newVote;
         }
     }
 
@@ -71,14 +115,40 @@ public class VoteServiceImpl implements VoteService {
         Review review = reviewRepository.findById(reviewId)
                 .orElseThrow(() -> new WeviewAPIException(HttpStatus.BAD_REQUEST, "Review not found!"));
 
+        if (review.getUser().getId().equals(user.getId())) {
+            throw new WeviewAPIException(HttpStatus.FORBIDDEN, "Cannot vote on own review!");
+        }
+
         Optional<Vote> existingVote = voteRepository.findByReview_IdAndUserId(review.getId(), user.getId());
         if (existingVote.isPresent()) {
             if (existingVote.get().getVoteType().equals(voteType)) {
                 voteRepository.delete(existingVote.get());
+                if (voteType.equals(VoteType.DOWNVOTE)) {
+                    // REMOVE DOWNVOTE
+                    userService.modifyPoints(review.getUser().getId(), 4);
+                } else if (voteType.equals(VoteType.UPVOTE)) {
+                    //REMOVE UPVOTE
+                    userService.modifyPoints(review.getUser().getId(), -4);
+                }
+
+                // Point change to current user
+                // Removing vote, so remove points
+                userService.modifyPoints(user.getId(), -1);
                 return null;
             } else {
                 existingVote.get().setVoteType(voteType);
-                return voteRepository.save(existingVote.get());
+                Vote vote = voteRepository.save(existingVote.get());
+
+                // Inverting votes
+                if (voteType.equals(VoteType.DOWNVOTE)) {
+                    // REMOVE UPVOTE ADD DOWNVOTE
+                    userService.modifyPoints(review.getUser().getId(), -8);
+                } else if (voteType.equals(VoteType.UPVOTE)) {
+                    // REMOVE DOWNVOTE ADD UPVOTE
+                    userService.modifyPoints(review.getUser().getId(), +8);
+                }
+                // Invert points so no changes in current user points
+                return vote;
             }
         } else {
             Vote vote = new Vote();
@@ -86,7 +156,19 @@ public class VoteServiceImpl implements VoteService {
             vote.setReview(review);
             vote.setVoteType(voteType);
             vote.setUser(user);
-            return voteRepository.save(vote);
+
+            Vote newVote = voteRepository.save(vote);
+            if (voteType.equals(VoteType.DOWNVOTE)) {
+                // ADD DOWNVOTE
+                userService.modifyPoints(review.getUser().getId(), -4);
+            } else if (voteType.equals(VoteType.UPVOTE)) {
+                // ADD UPVOTE
+                userService.modifyPoints(review.getUser().getId(), 4);
+            }
+
+            // +1 for action
+            userService.modifyPoints(user.getId(), 1);
+            return newVote;
         }
     }
